59 full app integration (#83)
* Optimize temporal workflow and update pocket config

Improve the workflow initialization and scheduling process in the morse-poc environment. Replace ThreadPoolExecutor with ProcessPoolExecutor for activity execution in python worker main scripts. Adjust pocket configurations to include lean peer dependency and new service chains for load testing. Also adjusted cpus limit and replica counts for services in docker compose files for sampler and evaluator.

* Add LookupTasks workflow and adapt existing workflows

In this commit, a new workflow, LookupTasks, was introduced to the task management process, which retrieves completed tasks for further operations. Modifications were also made to the existing workflows, including added logging and updated workflow timeouts. In addition, the code for handling database operations in MongoDB was updated.

* added multiple llm endpoints, minor fixs

* Modify initialization datasets and revise workflow constraints

Updated initialization datasets within temporal workflow configurations under morse-poc. Precisely adjusted task queue setup and execution timeouts in workflow constraints. Ensured consistent code formatting and style in python files under packages. Made additions to MongoDB configuration for handling task status and added index for results_data. Corrected identification reuse policy for workflows and removed unnecessary dependencies for certain docker services.

* Optimize temporal workflow and update pocket config

Improve the workflow initialization and scheduling process in the morse-poc environment. Replace ThreadPoolExecutor with ProcessPoolExecutor for activity execution in python worker main scripts. Adjust pocket configurations to include lean peer dependency and new service chains for load testing. Also adjusted cpus limit and replica counts for services in docker compose files for sampler and evaluator.

* Add LookupTasks workflow and adapt existing workflows

In this commit, a new workflow, LookupTasks, was introduced to the task management process, which retrieves completed tasks for further operations. Modifications were also made to the existing workflows, including added logging and updated workflow timeouts. In addition, the code for handling database operations in MongoDB was updated.

* added multiple llm endpoints, minor fixs

* Modify initialization datasets and revise workflow constraints

Updated initialization datasets within temporal workflow configurations under morse-poc. Precisely adjusted task queue setup and execution timeouts in workflow constraints. Ensured consistent code formatting and style in python files under packages. Made additions to MongoDB configuration for handling task status and added index for results_data. Corrected identification reuse policy for workflows and removed unnecessary dependencies for certain docker services.

* Added web app and a test-api in the meantime that @ramiro create the final API app

* mvp ok

---------

+++ apps/go/manager/records/task.go
@@ -359,7 +359,7 @@ type NumericalTaskRecord struct {
 }
 
 type ScoresSample struct {
-	Score float32 `bson:"score"`
+	Score float64 `bson:"score"`
 	ID    int     `bson:"id"`
 }
 
@@ -525,9 +525,9 @@ func (record *NumericalTaskRecord) ProcessData(l *zerolog.Logger) (err error) {
 		record.StdScore = 0
 		record.MedianScore = 0
 	} else if length == 1 {
-		record.MeanScore = record.ScoresSamples[record.CircBuffer.Indexes.Start].Score
+		record.MeanScore = float32(record.ScoresSamples[record.CircBuffer.Indexes.Start].Score)
 		record.StdScore = 0
-		record.MedianScore = record.ScoresSamples[record.CircBuffer.Indexes.Start].Score
+		record.MedianScore = float32(record.ScoresSamples[record.CircBuffer.Indexes.Start].Score)
 	} else {
 		// Calculate the mean
 		record.MeanScore = float32(stat.Mean(auxData, nil))
